Fixed more arithmetic overflow warnings with type casts
* Performed casts on getNumberOfComponents() to size_t for calculations.  This method should probably return size_t instead of a signed int if a negative component dimension is not possible.
* Changed DynamicTableData::ExpandData parameters nRows and nCols from signed int to size_t.  Negative values were not handled, and the values are more descriptive with their current type.  size_t should be prefered over int when it comes to dimensions unless there is a good reason to include negative values as a valid option.
* Added a hsize_t cast to H5DataArrayWriter when calculating rank size in H5DataArrayWriter.hpp  By casting the first value, arithmetic operations progress it forward while preventing int32_t-related overflow.
* Added casts to int64_t when faceId values to local variables in GeometryMath.cpp.  These variables were used for referencing vertex pointer indices where int64_t values were expected

+++ Source/SIMPLib/CoreFilters/CreateFeatureArrayFromElementArray.cpp
@@ -173,7 +173,7 @@ template <typename T> IDataArray::Pointer copyCellData(AbstractFilter* filter, I
   T* fPtr = feature->getPointer(0);
   T* cPtr = cell->getPointer(0);
 
-  int32_t numComp = cell->getNumberOfComponents();
+  size_t numComp = static_cast<size_t>(cell->getNumberOfComponents());
   int32_t featureIdx = 0;
 
   size_t cells = inputData->getNumberOfTuples();
